+++ src/Data/CreateParams.php
@@ -22,6 +22,7 @@ use Upmind\ProvisionBase\Provider\DataSet\Rules;
  * @property-read boolean|null $as_reseller Whether or not the new account should have reseller privileges
  * @property-read ResellerOptionParams|null $reseller_options Additional options for resellers
  * @property-read string|null $custom_ip Custom IP address for the new account
+ * @property-read string|null $location Location of the account
  */
 class CreateParams extends DataSet
 {
@@ -40,6 +41,7 @@ class CreateParams extends DataSet
             'as_reseller' => ['boolean'],
             'reseller_options' => [ResellerOptionParams::class],
             'custom_ip' => ['ip'],
+            'location' => ['nullable', 'string']
         ]);
     }
 }