+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the package will be documented in this file.
 
+## TBC
+
+- Add optional `location` parameter value to `CreateParams` and return value to `AccountInfo`
+  - Add logic to Enhance to select server group based on `location` parameter and return group name in `AccountInfo`
+  - Add logic to 20i to select data centre based on `location` parameter and return zone name in `AccountInfo`
+
 ## [v6.5.2](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.5.2) - 2024-01-08
 
 - Update DirectAdmin add `ip_status` configuration value to specify IP status type